Update TodoItem ListId type and improve EF configurations
Changed TodoItem.ListId from int to long for consistency or scalability. Enhanced Entity Framework configurations for TodoItem and TodoList entities by adding navigation property setup, foreign key constraints, and XML documentation comments.

+++ WebApplicationTemplate/src/Domain/Entities/TodoItem.cs
@@ -13,7 +13,7 @@ public class TodoItem : BaseAuditableEntity
     /// <summary>
     /// Gets or sets the identifier of the to-do list to which this item belongs.
     /// </summary>
-    public int ListId { get; set; }
+    public long ListId { get; set; }
 
     /// <summary>
     /// Gets or sets the title of the to-do item.
